add: API 서버 Open API 형식으로 응답 변경 및 swgger json으로 서버 타입 유형 생성 기능 추가

+++ apps/api/src/routes/auth/services/auth.service.ts
@@ -1,21 +1,31 @@
-import { HttpStatus, Injectable } from "@nestjs/common";
-import { JwtService } from "@nestjs/jwt";
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from "@nestjs/common";
 import { subMilliseconds } from "date-fns";
 
 import { HttpResultStatus } from "@template/sdk";
 
-import type { JwtPayload } from "../strategies/jwt.auth.strategy";
-import { EnvironmentService } from "../../../integrations/environment/environment.service";
 import { LoggerService } from "../../../integrations/logger/logger.service";
 import { PrismaService } from "../../../integrations/prisma/prisma.service";
 import { AppTokenType } from "../../../libs/constants";
-import { assertHttpError, isHttpError } from "../../../libs/error";
 import { UsersService } from "../../../routes/users/services/users.service";
+import { AuthResponseDto } from "../../../shared/dtos/response/auth/auth-response.dto";
 import { RefreshTokenDTO } from "../dto/refresh-token.dto";
 import { SigninDTO } from "../dto/signin.dto";
 import { SignoutDTO } from "../dto/signout.dto";
 import { SignupDTO } from "../dto/signup.dto";
 import { VerifyTokenDTO } from "../dto/verify-token.dto";
+import {
+  AuthAlreadyExistEmailErrorCode,
+  AuthErrorDefine,
+  AuthIncorrectPasswordErrorCode,
+  AuthNotExistEmailErrorCode,
+  AuthNotExistUserErrorCode,
+  AuthTokenNotExistErrorCode,
+} from "../errors";
 import { PasswordService } from "./password.service";
 import { TokenService } from "./token.service";
 
@@ -40,13 +50,11 @@ export class AuthService {
   private _contextName = "auth - service";
 
   constructor(
-    private readonly env: EnvironmentService,
     private readonly prisma: PrismaService,
     private readonly logger: LoggerService,
     private readonly token: TokenService,
     private readonly user: UsersService,
     private readonly password: PasswordService,
-    private readonly jwt: JwtService,
   ) {}
 
   /**
@@ -54,51 +62,19 @@ export class AuthService {
    * @param {VerifyTokenDTO} input
    */
   async verifyToken(input: VerifyTokenDTO) {
-    let jwtDto: JwtPayload;
-    try {
-      jwtDto = await this.jwt.verifyAsync<JwtPayload>(input.token, {
-        secret: this.env.getAccessTokenSecret(),
-      });
-    } catch {
-      assertHttpError(
-        true,
-        {
-          resultCode: HttpResultStatus.TOKEN_EXPIRED,
-          message: "Unauthorized",
-          result: null,
-        },
-        "Unauthorized",
-        HttpStatus.UNAUTHORIZED,
-      );
+    const jwtDto = await this.token.getAccessTokenPayload(input.token);
+    const token = await this.token.findByTokenId(jwtDto.jti);
+    if (!token) {
+      throw new NotFoundException(AuthErrorDefine[AuthTokenNotExistErrorCode]);
     }
 
-    assertHttpError(
-      !jwtDto || (jwtDto && !jwtDto.sub),
-      {
-        resultCode: HttpResultStatus.TOKEN_EXPIRED,
-        message: "Unauthorized",
-        result: null,
-      },
-      "Unauthorized",
-      HttpStatus.UNAUTHORIZED,
-    );
-
     const user = await this.user.checkUserById(jwtDto.sub);
-    assertHttpError(
-      !user,
-      {
-        resultCode: HttpResultStatus.NOT_EXIST,
-        message: "user not found",
-        result: null,
-      },
-      "Not Found",
-      HttpStatus.NOT_FOUND,
-    );
+    if (!user) {
+      throw new NotFoundException(AuthErrorDefine[AuthNotExistUserErrorCode]);
+    }
 
     return {
       resultCode: HttpResultStatus.OK,
-      message: null,
-      error: null,
       result: true,
     };
   }
@@ -107,98 +83,62 @@ export class AuthService {
    * @description Refresh Handler
    * @param {RefreshTokenDTO} input
    */
-  async refresh(input: RefreshTokenDTO) {
+  async refresh(input: RefreshTokenDTO): Promise<{
+    resultCode: HttpResultStatus;
+    data: AuthResponseDto;
+  }> {
     const jwtDto = await this.token.getRefreshTokenPayload(input.refreshToken);
-
     const token = await this.token.findByTokenId(jwtDto.jti);
-    assertHttpError(
-      !token,
-      {
-        resultCode: HttpResultStatus.TOKEN_EXPIRED,
-        message: "token not found",
-        result: null,
-      },
-      "Not Found",
-      HttpStatus.UNAUTHORIZED,
+    if (!token) {
+      throw new NotFoundException(AuthErrorDefine[AuthTokenNotExistErrorCode]);
+    }
+
+    const user = await this.user.getUserById(jwtDto.sub);
+    if (!user) {
+      throw new NotFoundException(AuthErrorDefine[AuthNotExistUserErrorCode]);
+    }
+
+    // 현재 발급한 token의 만료일 이전에 만료된 token을 삭제
+    await this.token.deleteByExpiresAtTokens({
+      userId: user.id,
+      expiresAt: subMilliseconds(token.expires.getTime(), 1000),
+      tokenType: AppTokenType.RefreshToken,
+    });
+
+    const accessToken = this.token.generateAccessToken(user.id);
+    const refreshToken = await this.token.generateRefreshToken(
+      user.id,
+      accessToken.token,
     );
 
-    try {
-      const user = await this.user.getUserById(jwtDto.sub);
-      assertHttpError(
-        !user,
-        {
-          resultCode: HttpResultStatus.NOT_EXIST,
-          message: "user not found",
-          result: null,
+    return {
+      resultCode: HttpResultStatus.OK,
+      data: {
+        id: user.id,
+        email: user.email,
+        name: user.name,
+        image: user.image,
+        tokens: {
+          accessToken,
+          refreshToken,
         },
-        "Not Found",
-        HttpStatus.NOT_FOUND,
-      );
-
-      // 현재 발급한 token의 만료일 이전에 만료된 token을 삭제
-      await this.token.deleteByExpiresAtTokens({
-        userId: user.id,
-        expiresAt: subMilliseconds(token.expires.getTime(), 1000),
-        tokenType: AppTokenType.RefreshToken,
-      });
-
-      const accessToken = this.token.generateAccessToken(user.id);
-      const refreshToken = await this.token.generateRefreshToken(
-        user.id,
-        accessToken.token,
-      );
-
-      return {
-        resultCode: HttpResultStatus.OK,
-        message: null,
-        error: null,
-        result: {
-          id: user.id,
-          email: user.email,
-          name: user.name,
-          image: user.image,
-          tokens: {
-            accessToken,
-            refreshToken,
-          },
-        },
-      };
-    } catch (error) {
-      if (isHttpError(error)) {
-        try {
-          await this.token.deleteByTokenId(jwtDto.jti);
-        } catch (error) {
-          // Nothing to do
-          this.logger.error(error, this._contextName);
-        }
-      }
-
-      throw error;
-    }
+      },
+    };
   }
 
   /**
    * @description Signin Handler
    * @param {SigninDTO} input
    */
-  async signin(input: SigninDTO) {
+  async signin(input: SigninDTO): Promise<{
+    resultCode: HttpResultStatus;
+    data: AuthResponseDto;
+  }> {
     const user = await this.user.getInternalUserByEmail(input.email);
 
-    assertHttpError(
-      !user,
-      {
-        resultCode: HttpResultStatus.NOT_EXIST,
-        message: {
-          email: {
-            message: "가입되지 않은 이메일입니다.",
-          },
-        },
-        error: "email",
-        result: null,
-      },
-      "Not Found",
-      HttpStatus.NOT_FOUND,
-    );
+    if (!user) {
+      throw new NotFoundException(AuthErrorDefine[AuthNotExistEmailErrorCode]);
+    }
 
     const isMatch = await this.password.compare(
       input.password,
@@ -206,21 +146,11 @@ export class AuthService {
       user.Password.hash,
     );
 
-    assertHttpError(
-      !isMatch,
-      {
-        resultCode: HttpResultStatus.INCORRECT_PASSWORD,
-        message: {
-          password: {
-            message: "비밀번호가 일치하지 않습니다.",
-          },
-        },
-        error: "password",
-        result: null,
-      },
-      "Unauthorized",
-      HttpStatus.UNAUTHORIZED,
-    );
+    if (!isMatch) {
+      throw new UnauthorizedException(
+        AuthErrorDefine[AuthIncorrectPasswordErrorCode],
+      );
+    }
 
     const accessToken = this.token.generateAccessToken(user.id);
     const refreshToken = await this.token.generateRefreshToken(
@@ -237,9 +167,7 @@ export class AuthService {
 
     return {
       resultCode: HttpResultStatus.OK,
-      message: null,
-      error: null,
-      result: {
+      data: {
         id: user.id,
         email: user.email,
         name: user.name,
@@ -256,32 +184,26 @@ export class AuthService {
    * @description Signup Handler
    * @param {SignupDTO} input
    */
-  async signup(input: SignupDTO) {
+  async signup(input: SignupDTO): Promise<{
+    resultCode: HttpResultStatus;
+    data: AuthResponseDto;
+  }> {
     const user = await this.user.checkUserByEmail(input.email);
 
-    assertHttpError(
-      !!user,
-      {
-        resultCode: HttpResultStatus.NOT_EXIST,
-        message: {
-          email: {
-            message: "이미 가입된 이메일입니다.",
-          },
-        },
-        error: "email",
-        result: null,
-      },
-      "Bad Request",
-      HttpStatus.BAD_REQUEST,
-    );
-
-    const { hashedPassword, salt } = await this.password.hashPassword(
-      input.password,
-    );
-
-    const emailSplit = input.email.split("@").at(0) ?? "username";
+    if (user) {
+      throw new BadRequestException(
+        AuthErrorDefine[AuthAlreadyExistEmailErrorCode],
+        "중복 회원가입 요청할때 발생하는 오류",
+      );
+    }
 
     return await this.prisma.$transaction(async (tx) => {
+      const { hashedPassword, salt } = await this.password.hashPassword(
+        input.password,
+      );
+
+      const emailSplit = input.email.split("@").at(0) ?? "username";
+
       const user = await this.user.createUser(
         {
           email: input.email,
@@ -292,7 +214,9 @@ export class AuthService {
         tx,
       );
 
+      // 인증 토큰 생성
       const accessToken = this.token.generateAccessToken(user.id);
+      // 리프레시 토큰 생성
       const refreshToken = await this.token.generateRefreshToken(
         user.id,
         accessToken.token,
@@ -305,6 +229,7 @@ export class AuthService {
         1000,
       );
 
+      // 현재 발급한 token의 만료일 이전에 만료된 token을 삭제
       await this.token.deleteByExpiresAtTokens(
         {
           userId: user.id,
@@ -316,9 +241,7 @@ export class AuthService {
 
       return {
         resultCode: HttpResultStatus.OK,
-        message: null,
-        error: null,
-        result: {
+        data: {
           id: user.id,
           email: user.email,
           name: user.name,
@@ -336,28 +259,26 @@ export class AuthService {
    * @description Signout Handler
    * @param {SignoutDTO} input
    */
-  async signout(input: SignoutDTO) {
+  async signout(input: SignoutDTO): Promise<{
+    resultCode: HttpResultStatus;
+    data: boolean;
+  }> {
     const jwtDto = await this.token.getAccessTokenPayload(input.accessToken);
+    const token = await this.token.findByTokenId(jwtDto.jti);
+    if (!token) {
+      throw new NotFoundException(AuthErrorDefine[AuthTokenNotExistErrorCode]);
+    }
 
     const user = await this.user.checkUserById(jwtDto.sub);
-    assertHttpError(
-      !user,
-      {
-        resultCode: HttpResultStatus.NOT_EXIST,
-        message: "user not found",
-        result: null,
-      },
-      "Not Found",
-      HttpStatus.NOT_FOUND,
-    );
+    if (!user) {
+      throw new NotFoundException(AuthErrorDefine[AuthNotExistUserErrorCode]);
+    }
 
     await this.token.deleteByAccessToken(input.accessToken);
 
     return {
       resultCode: HttpResultStatus.OK,
-      message: null,
-      error: null,
-      result: true,
+      data: true,
     };
   }
 }
